requirements.txt
streamlit
ultralytics
opencv-python
pandas
torch
openai
python-dotenv
flask
